+++ R/tdm.R
@@ -560,3 +560,60 @@ t.DocumentTermMatrix <- function(x, ...) {
     names(a) <- c("i", "j", "v", "nrow", "ncol", "dimnames")
     a
 }
+
+#' Apply a tm Corpus as a qdap Dataframe
+#' 
+#' Apply a tm \code{\link[tm]{Corpus}} as a qdap dataframe.
+#' 
+#' @param qdapfun A qdap function that is usually used on 
+#' text.variable ~ grouping variable.
+#' @note \code{aply_ad_df} coherses to a dataframe with columns named `docs` and 
+#' the othe rnames `docs`.
+#' @export
+#' @examples
+#' library(tm)
+#' reut21578 <- system.file("texts", "crude", package = "tm")
+#' reuters <- Corpus(DirSource(reut21578),
+#'     readerControl = list(reader = readReut21578XML))
+#' 
+#' apply_as_df(reuters, word_stats)
+#' apply_as_df(reuters, formality)
+#' apply_as_df(reuters, word_list)
+#' apply_as_df(reuters, polarity)
+#' apply_as_df(reuters, Dissimilarity)
+#' apply_as_df(reuters, diversity)
+#' apply_as_df(reuters, pos_by)
+#' apply_as_df(reuters, flesch_kincaid)
+#' apply_as_df(reuters, trans_venn)
+#' apply_as_df(reuters, gantt_plot)
+#' apply_as_df(reuters, rank_freq_mplot)
+#' apply_as_df(reuters, termco, 
+#'     match.list = list(
+#'         oil = qcv(oil, Texas, crude), 
+#'         money = c("economic", "money")
+#'     ))
+#' plot(apply_as_df(reuters, termco, 
+#'     match.list = list(
+#'         oil = qcv(oil, Texas, crude), 
+#'         money = c("economic", "money")
+#'     ), elim.old = FALSE), values = TRUE, high="red")
+#' apply_as_df(reuters, word_cor, 
+#'     word = qcv(oil, Texas, crude, economic, money)
+#' )
+#' plot(apply_as_df(reuters, word_cor, 
+#'     word = qcv(oil, Texas, crude, economic, money)
+#' ))
+apply_as_df <- function(tm.corpus, qdapfun, ...) {
+
+    text <- doc <- tot <- NULL
+
+    dat <- sentSplit(tm_corpus2df(tm.corpus), "text")
+    if (any(unlist(formals(qdapfun)) %in% "tot")) {
+        with(dat, qdapfun(text.var = text, grouping.var = docs, tot = tot,  ...))
+    } else {
+        with(dat, qdapfun(text.var = text, grouping.var = docs, ...))
+    }
+
+}
+
+